remove toolkits (move to tools index), pass apify token as arg, clean up old tools.ts, separate Actor readme
