+++ .actor/actor.json
@@ -5,5 +5,6 @@
   "description": "Implementation of a Model Context Protocol (MCP) Server for Apify Actors that enables AI applications (and AI agents) to interact with Apify Actors",
   "version": "0.1",
   "input": "./input_schema.json",
+  "readme": "./ACTOR.md",
   "dockerfile": "./Dockerfile"
 }
