+++ src/actors.ts
@@ -1,9 +1,10 @@
 import { Ajv } from 'ajv';
+import type { ActorStoreList } from 'apify-client';
 import { ApifyClient } from 'apify-client';
 
 import { ACTOR_ADDITIONAL_INSTRUCTIONS, defaults, MAX_DESCRIPTION_LENGTH, ACTOR_README_MAX_LENGTH, ACTOR_ENUM_MAX_LENGTH } from './const.js';
 import { log } from './logger.js';
-import type { ActorDefinitionPruned, ActorDefinitionWithDesc, IActorInputSchema, ISchemaProperties, ToolWrap } from './types.js';
+import type { ActorStorePruned, PricingInfo, ActorDefinitionPruned, ActorDefinitionWithDesc, IActorInputSchema, ISchemaProperties, ToolWrap } from './types.js';
 
 export function actorNameToToolName(actorName: string): string {
     return actorName
@@ -348,4 +349,40 @@ export async function getActorsAsTools(actors: string[]): Promise<ToolWrap[]> {
         }
     }
     return tools;
+}
+
+function pruneActorStoreInfo(response: ActorStoreList): ActorStorePruned {
+    const stats = response.stats || {};
+    const pricingInfo = (response.currentPricingInfo || {}) as PricingInfo;
+    return {
+        id: response.id,
+        name: response.name?.toString() || '',
+        username: response.username?.toString() || '',
+        actorFullName: `${response.username}/${response.name}`,
+        title: response.title?.toString() || '',
+        description: response.description?.toString() || '',
+        stats: {
+            totalRuns: stats.totalRuns,
+            totalUsers30Days: stats.totalUsers30Days,
+            publicActorRunStats30Days: 'publicActorRunStats30Days' in stats
+                ? stats.publicActorRunStats30Days : {},
+        },
+        currentPricingInfo: {
+            pricingModel: pricingInfo.pricingModel?.toString() || '',
+            pricePerUnitUsd: pricingInfo?.pricePerUnitUsd ?? 0,
+            trialMinutes: pricingInfo?.trialMinutes ?? 0,
+        },
+        url: response.url?.toString() || '',
+        totalStars: 'totalStars' in response ? (response.totalStars as number) : null,
+    };
+}
+
+export async function searchActorsByKeywords(
+    search: string,
+    limit: number | undefined = undefined,
+    offset: number | undefined = undefined,
+): Promise<ActorStorePruned[] | null> {
+    const client = new ApifyClient({ token: process.env.APIFY_TOKEN });
+    const results = await client.store().list({ search, limit, offset });
+    return results.items.map((x) => pruneActorStoreInfo(x));
 }
