--- src/toolkits/index.ts
@@ -1,2 +1,0 @@
-export { getActorAutoLoadingTools } from './actor-auto-loading-tools.js';
-export { getActorDiscoveryTools } from './discovery-tools.js';
